refactor(scene): consolidate time_of_day into sky (move-only, byte-identical)
