+++ crates/pixtuoid-scene/src/pixel_painter/background/sky.rs
